backend/src/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pillow==10.0.0
numpy==1.24.3
scipy==1.10.1
opencv-python==4.8.0.76
transformers==4.33.2
torch==2.0.1
flask-mail==0.9.1
langchain-google-genai==0.0.5
cloudinary==1.34.0
pymongo==4.3.3
PyJWT==2.8.0